Migrate 'has-class' test
